removal of my own OneHotFrequentSubtreeFeatures classes.
It is interesting to note that the sklearn LinearRegression algorithm
is faster *and produces less error* when applied to the sparse OneHot
features produced by the sklearn class applied to the data generated
by my FeatureGeneratingTree class compared to the same alg applied to
dense dataset produced by my OneHotFeatureGeneratingTree class.

This is the case, although the data matrices contain identical data.